Moved inner functions into utils module
