+++ custom/ilsgateway/tanzania/reports/stock_on_hand.py
@@ -1,6 +1,5 @@
 from casexml.apps.stock.models import StockTransaction
-from datetime import timedelta, datetime, time
-from django.db.models import Q
+from datetime import timedelta, datetime
 from django.template.defaultfilters import floatformat
 from corehq.apps.commtrack.models import SQLProduct, StockState
 from corehq.apps.locations.models import SQLLocation
@@ -8,17 +7,18 @@ from corehq.apps.reports.datatables import DataTablesHeader, DataTablesColumn
 from corehq.apps.reports.filters.fixtures import AsyncLocationFilter
 from django.utils import html
 from custom.ilsgateway.filters import ProgramFilter, ILSDateFilter
-from custom.ilsgateway.models import SupplyPointStatusTypes, ProductAvailabilityData, \
-    OrganizationSummary, SupplyPointStatus, SupplyPointStatusValues
-from custom.ilsgateway.tanzania import ILSData, DetailsReport
+from custom.ilsgateway.models import ProductAvailabilityData, \
+    OrganizationSummary
 from custom.ilsgateway.tanzania.reports.facility_details import FacilityDetailsReport, InventoryHistoryData, \
     RegistrationData, RandRHistory, RecentMessages, Notes
 from custom.ilsgateway.tanzania.reports.mixins import ProductAvailabilitySummary, SohSubmissionData
 from django.utils.translation import ugettext as _
-from custom.ilsgateway.tanzania.reports.utils import link_format, format_percent, make_url
-from dimagi.utils.dates import get_business_day_of_month, get_day_of_month
+from custom.ilsgateway.tanzania.reports.utils import link_format, format_percent, make_url, get_hisp_resp_rate, \
+    get_last_reported, calculate_months_remaining
+from dimagi.utils.dates import get_day_of_month
 from dimagi.utils.decorators.memoized import memoized
 from django.db.models.aggregates import Avg, Max
+from custom.ilsgateway.tanzania import ILSData, DetailsReport
 
 
 def get_facilities(location, domain):
@@ -264,50 +264,6 @@ class DistrictSohPercentageTableData(ILSData):
         rows = []
         enddate = self.config['enddate']
 
-        def get_last_reported(supplypoint):
-
-            last_bd_of_the_month = get_business_day_of_month(enddate.year,
-                                                             enddate.month,
-                                                             -1)
-            st = StockTransaction.objects.filter(
-                case_id=supplypoint,
-                type='stockonhand',
-                report__date__lte=last_bd_of_the_month,
-                report__domain=self.config['domain']
-            ).order_by('-report__date')
-            last_of_last_month = datetime(enddate.year, enddate.month, 1) - timedelta(days=1)
-            last_bd_of_last_month = datetime.combine(get_business_day_of_month(last_of_last_month.year,
-                                                     last_of_last_month.month,
-                                                     -1), time())
-            if st:
-                sts = _reported_on_time(last_bd_of_last_month, st[0].report.date)
-                return sts, st[0].report.date.date()
-            else:
-                sts = OnTimeStates.NO_DATA
-                return sts, None
-
-        def get_hisp_resp_rate(location):
-            statuses = SupplyPointStatus.objects.filter(
-                location_id=location.location_id,
-                status_type=SupplyPointStatusTypes.SOH_FACILITY
-            )
-            if not statuses:
-                return None
-            status_month_years = set([(x.status_date.month, x.status_date.year) for x in statuses])
-            denom = len(status_month_years)
-            num = 0
-            for s in status_month_years:
-                f = statuses.filter(status_date__month=s[0], status_date__year=s[1]).filter(
-                    Q(status_value=SupplyPointStatusValues.SUBMITTED) |
-                    Q(status_value=SupplyPointStatusValues.NOT_SUBMITTED) |
-                    Q(status_value=SupplyPointStatusValues.RECEIVED) |
-                    Q(status_value=SupplyPointStatusValues.NOT_RECEIVED)
-                ).order_by("-status_date")
-                if f.count():
-                    num += 1
-
-            return float(num) / float(denom), num, denom
-
         if not self.config['products']:
             products = SQLProduct.objects.filter(domain=self.config['domain'], is_archived=False).order_by('code')
         else:
@@ -319,7 +275,8 @@ class DistrictSohPercentageTableData(ILSData):
             for loc in locations:
                 supply_point = loc.supply_point_id
 
-                status, last_reported = get_last_reported(supply_point)
+                status, last_reported = get_last_reported(supply_point, self.config['domain'], enddate)
+                status, last_reported = get_last_reported(supply_point, enddate, self.config['domain'])
                 hisp = get_hisp_resp_rate(loc)
 
                 url = make_url(FacilityDetailsReport, self.config['domain'],
@@ -353,13 +310,6 @@ class DistrictSohPercentageTableData(ILSData):
                         srs = None
 
                     if srs:
-                        def calculate_months_remaining(stock_state, quantity):
-                            consumption = stock_state.daily_consumption
-                            if consumption is not None and consumption > 0 and quantity is not None:
-                                return float(quantity) / float(30 * consumption)
-                            elif quantity == 0:
-                                return 0
-                            return None
                         try:
                             ss = StockState.objects.get(case_id=supply_point, product_id=product.product_id)
                             val = calculate_months_remaining(ss, srs.stock_on_hand)
